backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "boto3>=1.40.25",
    "fastapi-mail>=1.5.0",
    "fastapi[all]>=0.116.1",
    "passlib[argon2]>=1.7.4",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
